Resolve "Add new tests for the product controller"

See merge request cse2115/2020-2010/7-student-house-food-management/op26-sem51!28

+++ transactions/src/main/java/nl/tudelft/sem/transactions/controllers/ProductController.java
@@ -133,7 +133,7 @@ public class ProductController {
      */
     @DeleteMapping("/deleteProduct")
     public @ResponseBody
-    boolean deleteProduct(@RequestParam int productId) {
+    boolean deleteProduct(@RequestParam long productId) {
         try {
             return productRepository.deleteProductById(productId) != 0;
         } catch (Exception e) {
